+++ R/dataframe__frame.R
@@ -525,7 +525,7 @@ DataFrame_lazy = "use_extendr_wrapper"
 #' # ... but simply assigning df1 to df3 change the address anyway
 #' pl$mem_address(df1) == pl$mem_address(df3)
 DataFrame_clone = function() {
-  .pr$DataFrame$clone_see_me_macro(self)
+  .pr$DataFrame$clone_in_rust(self)
 }
 
 #' Get columns (as Series)
@@ -856,11 +856,6 @@ DataFrame_to_data_frame = function(...) {
   df[] = lapply(df, unAsIs)
   df
 }
-
-#' Alias for to_data_frame (backward compatibility)
-#' @return An R data.frame
-#' @noRd
-DataFrame_as_data_frame = DataFrame_to_data_frame
 
 
 #' Return Polars DataFrame as a list of vectors