+++ R/group_by.R
@@ -285,20 +285,20 @@ GroupBy_null_count = function() {
   self$agg(pl$all()$null_count())
 }
 
-#' convert to data.frame
-#'
-#' @param ... not used
-#'
-#' @return R data.frame
-#'
-#' @examples pl$DataFrame(iris)$to_data_frame() # R-polars back and forth
-GroupBy_to_data_frame = function(...) {
-  class(self) = "DataFrame"
-  self$to_data_frame(...)
-}
 
-# TODO REMOVE_AT_BREAKING_CHANGE
-#' Alias to GroupBy_to_data_frame (backward compatibility)
-#' @return R data.frame
-#' @noRd
-GroupBy_as_data_frame = GroupBy_to_data_frame
+#' GroupBy_ungroup
+#'
+#' Revert the group by operation.
+#' @return [DataFrame][DataFrame_class]
+#' @examples
+#' gb = pl$DataFrame(mtcars)$group_by("cyl")
+#' gb
+#'
+#' gb$ungroup()
+#' @export
+GroupBy_ungroup = function() {
+  self = .pr$DataFrame$clone_in_rust(self)
+  class(self) = "DataFrame"
+  attr(self, "private") = NULL
+  self
+}