+++ src/rust/src/lazy/dsl.rs
@@ -1,13 +1,12 @@
 use crate::concurrent::RFnSignature;
-use crate::rdatatype::literal_to_any_value;
-use crate::rdatatype::new_rank_method;
-use crate::rdatatype::new_rolling_cov_options;
-use crate::rdatatype::parse_fill_null_strategy;
-use crate::rdatatype::robj_to_timeunit;
-use crate::rdatatype::{DataTypeVector, RPolarsDataType};
+use crate::rdatatype::{
+    literal_to_any_value, new_rank_method, new_rolling_cov_options, parse_fill_null_strategy,
+, robj_to_timeunit, DataTypeVector, RPolarsDataType,
+};
 use crate::robj_to;
-use crate::rpolarserr::polars_to_rpolars_err;
-use crate::rpolarserr::{rerr, rpolars_to_polars_err, RResult, Rctx, WithRctx};
+use crate::rpolarserr::{
+    polars_to_rpolars_err, rerr, rpolars_to_polars_err, RResult, Rctx, WithRctx,
+};
 use crate::series::Series;
 use crate::utils::extendr_concurrent::{ParRObj, ThreadCom};
 use crate::utils::extendr_helpers::robj_inherits;
@@ -18,11 +17,14 @@ use crate::utils::{try_f64_into_i64, try_f64_into_u32, try_f64_into_usize};
 use crate::CONFIG;
 use extendr_api::{extendr, prelude::*, rprintln, Deref, DerefMut, Rinternals};
 use pl::PolarsError as pl_error;
-use pl::{BinaryNameSpaceImpl, DurationMethods, IntoSeries, TemporalMethods, Utf8NameSpaceImpl};
+use pl::{
+    BinaryNameSpaceImpl, Duration, DurationMethods, IntoSeries, RollingGroupOptions,
+    TemporalMethods, Utf8NameSpaceImpl,
+};
 use polars::lazy::dsl;
 use polars::prelude as pl;
 use polars::prelude::SortOptions;
-use std::ops::{Add, Div, Mul, Sub};
+use std::ops::{Add, Div, Mul, Rem, Sub};
 use std::result::Result;
 pub type NameGenerator = pl::Arc<dyn Fn(usize) -> String + Send + Sync>;
 use crate::rdatatype::robjs_to_ewm_options;
@@ -174,7 +176,7 @@ impl Expr {
         .when("constructing polars literal from Robj")
     }
 
-    //expr binary comparisons
+    //comparison
     pub fn gt(&self, other: Robj) -> RResult<Self> {
         Ok(self.0.clone().gt(robj_to!(PLExpr, other)?).into())
     }
@@ -195,11 +197,19 @@ impl Expr {
         Ok(self.0.clone().neq(robj_to!(PLExpr, other)?).into())
     }
 
+    pub fn neq_missing(&self, other: Robj) -> RResult<Self> {
+        Ok(self.0.clone().neq_missing(robj_to!(PLExpr, other)?).into())
+    }
+
     pub fn eq(&self, other: Robj) -> RResult<Self> {
         Ok(self.0.clone().eq(robj_to!(PLExpr, other)?).into())
     }
 
-    //logical operators
+    pub fn eq_missing(&self, other: Robj) -> RResult<Self> {
+        Ok(self.0.clone().eq_missing(robj_to!(PLExpr, other)?).into())
+    }
+
+    //conjunction
     fn and(&self, other: Robj) -> RResult<Self> {
         Ok(self.0.clone().and(robj_to!(PLExpr, other)?).into())
     }
@@ -208,12 +218,9 @@ impl Expr {
         Ok(self.0.clone().or(robj_to!(PLExpr, other)?).into())
     }
 
+    //binary
     fn xor(&self, other: Robj) -> RResult<Self> {
         Ok(self.0.clone().xor(robj_to!(PLExpr, other)?).into())
-    }
-
-    fn is_in(&self, other: Robj) -> RResult<Self> {
-        Ok(self.0.clone().is_in(robj_to!(PLExpr, other)?).into())
     }
 
     //any not translated expr from expr/expr.py
@@ -399,10 +406,6 @@ impl Expr {
 
     pub fn arg_unique(&self) -> Self {
         self.clone().0.arg_unique().into()
-    }
-
-    pub fn is_duplicated(&self) -> Self {
-        self.clone().0.is_duplicated().into()
     }
 
     pub fn quantile(&self, quantile: Robj, interpolation: Robj) -> RResult<Self> {
@@ -1350,76 +1353,79 @@ impl Expr {
         ))
     }
 
-    pub fn duration_days(&self) -> Self {
-        self.0
+    pub fn dt_total_days(&self) -> RResult<Self> {
+        Ok(self
+            .0
             .clone()
             .map(
                 |s| Ok(Some(s.duration()?.days().into_series())),
                 pl::GetOutput::from_type(pl::DataType::Int64),
             )
-            .into()
+            .into())
     }
-    pub fn duration_hours(&self) -> Self {
-        self.0
+    pub fn dt_total_hours(&self) -> RResult<Self> {
+        Ok(self
+            .0
             .clone()
             .map(
                 |s| Ok(Some(s.duration()?.hours().into_series())),
                 pl::GetOutput::from_type(pl::DataType::Int64),
             )
-            .into()
+            .into())
     }
-    pub fn duration_minutes(&self) -> Self {
-        self.0
+    pub fn dt_total_minutes(&self) -> RResult<Self> {
+        Ok(self
+            .0
             .clone()
             .map(
                 |s| Ok(Some(s.duration()?.minutes().into_series())),
                 pl::GetOutput::from_type(pl::DataType::Int64),
             )
-            .into()
+            .into())
     }
-    pub fn duration_seconds(&self) -> Self {
-        self.0
+    pub fn dt_total_seconds(&self) -> RResult<Self> {
+        Ok(self
+            .0
             .clone()
             .map(
                 |s| Ok(Some(s.duration()?.seconds().into_series())),
                 pl::GetOutput::from_type(pl::DataType::Int64),
             )
-            .into()
+            .into())
     }
-    pub fn duration_nanoseconds(&self) -> Self {
-        self.0
-            .clone()
-            .map(
-                |s| Ok(Some(s.duration()?.nanoseconds().into_series())),
-                pl::GetOutput::from_type(pl::DataType::Int64),
-            )
-            .into()
-    }
-    pub fn duration_microseconds(&self) -> Self {
-        self.0
-            .clone()
-            .map(
-                |s| Ok(Some(s.duration()?.microseconds().into_series())),
-                pl::GetOutput::from_type(pl::DataType::Int64),
-            )
-            .into()
-    }
-    pub fn duration_milliseconds(&self) -> Self {
-        self.0
+    pub fn dt_total_milliseconds(&self) -> RResult<Self> {
+        Ok(self
+            .0
             .clone()
             .map(
                 |s| Ok(Some(s.duration()?.milliseconds().into_series())),
                 pl::GetOutput::from_type(pl::DataType::Int64),
             )
-            .into()
+            .into())
+    }
+    pub fn dt_total_microseconds(&self) -> RResult<Self> {
+        Ok(self
+            .0
+            .clone()
+            .map(
+                |s| Ok(Some(s.duration()?.microseconds().into_series())),
+                pl::GetOutput::from_type(pl::DataType::Int64),
+            )
+            .into())
+    }
+    pub fn dt_total_nanoseconds(&self) -> RResult<Self> {
+        Ok(self
+            .0
+            .clone()
+            .map(
+                |s| Ok(Some(s.duration()?.nanoseconds().into_series())),
+                pl::GetOutput::from_type(pl::DataType::Int64),
+            )
+            .into())
     }
 
     pub fn dt_offset_by(&self, by: Robj) -> RResult<Self> {
         Ok(self.clone().0.dt().offset_by(robj_to!(PLExpr, by)?).into())
-    }
-
-    pub fn pow(&self, exponent: Robj) -> RResult<Self> {
-        Ok(self.0.clone().pow(robj_to!(PLExpr, exponent)?).into())
     }
 
     pub fn repeat_by(&self, by: &Expr) -> Self {
@@ -1452,30 +1458,6 @@ impl Expr {
 
     pub fn alias(&self, s: &str) -> Self {
         self.0.clone().alias(s).into()
-    }
-
-    pub fn is_null(&self) -> Self {
-        self.0.clone().is_null().into()
-    }
-
-    pub fn is_not_null(&self) -> Self {
-        self.0.clone().is_not_null().into()
-    }
-
-    pub fn is_finite(&self) -> Self {
-        self.0.clone().is_finite().into()
-    }
-
-    pub fn is_infinite(&self) -> Self {
-        self.0.clone().is_infinite().into()
-    }
-
-    pub fn is_nan(&self) -> Self {
-        self.0.clone().is_nan().into()
-    }
-
-    pub fn is_not_nan(&self) -> Self {
-        self.0.clone().is_not_nan().into()
     }
 
     pub fn drop_nulls(&self) -> Self {
@@ -1559,11 +1541,59 @@ impl Expr {
         self.0.clone().agg_groups().into()
     }
 
+    // boolean
+
     pub fn all(&self, drop_nulls: Robj) -> RResult<Self> {
         Ok(self.0.clone().all(robj_to!(bool, drop_nulls)?).into())
     }
     pub fn any(&self, drop_nulls: Robj) -> RResult<Self> {
         Ok(self.0.clone().any(robj_to!(bool, drop_nulls)?).into())
+    }
+
+    // TODO: is_between
+
+    pub fn is_duplicated(&self) -> Self {
+        self.clone().0.is_duplicated().into()
+    }
+
+    pub fn is_finite(&self) -> Self {
+        self.0.clone().is_finite().into()
+    }
+
+    // TODO: rename to is_first_distinct
+    pub fn is_first(&self) -> Self {
+        self.clone().0.is_first_distinct().into()
+    }
+
+    fn is_in(&self, other: Robj) -> RResult<Self> {
+        Ok(self.0.clone().is_in(robj_to!(PLExpr, other)?).into())
+    }
+
+    pub fn is_infinite(&self) -> Self {
+        self.0.clone().is_infinite().into()
+    }
+
+    // TODO: is_last_distinct
+
+    pub fn is_nan(&self) -> Self {
+        self.0.clone().is_nan().into()
+    }
+    pub fn is_not_null(&self) -> Self {
+        self.0.clone().is_not_null().into()
+    }
+
+    pub fn is_not_nan(&self) -> Self {
+        self.0.clone().is_not_nan().into()
+    }
+    pub fn is_null(&self) -> Self {
+        self.0.clone().is_null().into()
+    }
+
+    pub fn is_unique(&self) -> Self {
+        self.0.clone().is_unique().into()
+    }
+    pub fn not_(&self) -> Self {
+        self.0.clone().not().into()
     }
 
     pub fn count(&self) -> Self {
@@ -1593,27 +1623,34 @@ impl Expr {
             .into()
     }
 
-    //binary arithmetic expressions
+    //numeric
+
     pub fn add(&self, other: Robj) -> RResult<Self> {
         Ok(self.0.clone().add(robj_to!(PLExpr, other)?).into())
     }
 
-    //binary arithmetic expressions
-    pub fn sub(&self, other: Robj) -> RResult<Self> {
-        Ok(self.0.clone().sub(robj_to!(PLExpr, other)?).into())
+    pub fn floor_div(&self, other: Robj) -> RResult<Self> {
+        Ok(self.0.clone().floor_div(robj_to!(PLExpr, other)?).into())
+    }
+
+    pub fn rem(&self, other: Robj) -> RResult<Self> {
+        Ok(self.0.clone().rem(robj_to!(PLExpr, other)?).into())
     }
 
     pub fn mul(&self, other: Robj) -> RResult<Self> {
         Ok(self.0.clone().mul(robj_to!(PLExpr, other)?).into())
     }
 
+    pub fn sub(&self, other: Robj) -> RResult<Self> {
+        Ok(self.0.clone().sub(robj_to!(PLExpr, other)?).into())
+    }
+
     pub fn div(&self, other: Robj) -> RResult<Self> {
         Ok(self.0.clone().div(robj_to!(PLExpr, other)?).into())
     }
 
-    //unary
-    pub fn not_(&self) -> Self {
-        self.0.clone().not().into()
+    pub fn pow(&self, exponent: Robj) -> RResult<Self> {
+        Ok(self.0.clone().pow(robj_to!(PLExpr, exponent)?).into())
     }
 
     //expr      "funnies"
@@ -1714,32 +1751,32 @@ impl Expr {
         self.0.clone().apply(rbgfunc, output_map).into()
     }
 
-    pub fn is_unique(&self) -> Self {
-        self.0.clone().is_unique().into()
-    }
-
     pub fn approx_n_unique(&self) -> Self {
         self.clone().0.approx_n_unique().into()
     }
 
-    pub fn is_first(&self) -> Self {
-        self.clone().0.is_first_distinct().into()
-    }
-
     // name methods
-    pub fn name_keep(&self) -> Self {
-        self.0.clone().name().keep().into()
+    pub fn name_keep(&self) -> RResult<Self> {
+        Ok(self.0.clone().name().keep().into())
     }
 
-    fn name_suffix(&self, suffix: String) -> Self {
-        self.0.clone().name().suffix(suffix.as_str()).into()
+    fn name_suffix(&self, suffix: String) -> RResult<Self> {
+        Ok(self.0.clone().name().suffix(suffix.as_str()).into())
     }
 
-    fn name_prefix(&self, prefix: String) -> Self {
-        self.0.clone().name().prefix(prefix.as_str()).into()
+    fn name_prefix(&self, prefix: String) -> RResult<Self> {
+        Ok(self.0.clone().name().prefix(prefix.as_str()).into())
     }
 
-    pub fn name_map(&self, lambda: Robj) -> Self {
+    fn name_to_lowercase(&self) -> RResult<Self> {
+        Ok(self.0.clone().name().to_lowercase().into())
+    }
+
+    fn name_to_uppercase(&self) -> RResult<Self> {
+        Ok(self.0.clone().name().to_uppercase().into())
+    }
+
+    pub fn name_map(&self, lambda: Robj) -> RResult<Self> {
         //find a way not to push lambda everytime to main thread handler
         //safety only accessed in main thread, can be temp owned by other threads
         let probj = ParRObj(lambda);
@@ -1787,7 +1824,7 @@ impl Expr {
                 .map(|str| str.to_string())
         };
 
-        self.clone().0.name().map(f).into()
+        Ok(self.clone().0.name().map(f).into())
     }
 
     //string methods
@@ -2328,6 +2365,31 @@ impl Expr {
             new_rolling_cov_options(window_size, min_periods, ddof)?,
         )
         .into())
+    }
+
+    pub fn rolling(
+        &self,
+        index_column: Robj,
+        period: Robj,
+        offset: Robj,
+        closed: Robj,
+        check_sorted: Robj,
+    ) -> RResult<Self> {
+        let index_column = robj_to!(String, index_column)?.into();
+        let period = Duration::parse(robj_to!(str, period)?);
+        let offset = Duration::parse(robj_to!(str, offset)?);
+        let closed_window = robj_to!(ClosedWindow, closed)?;
+        let check_sorted = robj_to!(bool, check_sorted)?;
+
+        let options = RollingGroupOptions {
+            index_column,
+            period,
+            offset,
+            closed_window,
+            check_sorted,
+        };
+
+        Ok(self.0.clone().rolling(options).into())
     }
 }